Ensure SLURM failures are properly propagated.
The previous code ran all scripts at the `cleanup_all_variables` step
which is the very last step of the `run.sh` and always executed.
+ While convenient, this step doesn't allow to propagate exit
codes (exit 1 at this step results in pipeline success! this is not
documented anywhere).
+ In the new format, everything up to the `step_script` step (the main
one containing before_script + build_script) is run in a SLURM job.
+ Starting from `after_script` (if provided by the user), all subsequent
steps are executed directly on the login node (without SLURM). It is
assumed that all performance critical steps happen before.